# cm.domeni.authentis-users/openapi/main.yaml
---
openapi: "3.0.3"
info:
  title: "cm.domeni.authentis-users"
  description: "Demo use open APi for generate REST API endpoints"
  version: "0.0.1"
servers:
- url: "https"
- url: "http"
paths:
  /demo:
    get:
      tags:
      - "Demo"
      summary: "Demo Summary"
      description: "Demo Description"
      operationId: "fetchAllDemo"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/DemoDTO"
        "401":
          description: "Unauthorized"
    post:
      tags:
      - "Demo"
      summary: "Create a new Demo"
      description: "Creates a new Demo with the provided data"
      operationId: "createDemo"
      security:
      - bearerAuth: []
      requestBody:
        description: "Demo object to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoDTO"
      responses:
        "201":
          description: "Demo created successfully"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          description: "Invalid input"
        "401":
          description: "Unauthorized"
  /register:
    post:
      tags:
      - "User"
      summary: "register user"
      description: "register user"
      operationId: "register"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: "string"
                format: "uuid"
        "400":
          description: "Bad request"
        "409":
          description: "this username already exist"
  /user:
    get:
      tags:
      - "User"
      summary: "get user info"
      description: "get user info including id, username, and all roles"
      operationId: "fetchAllUsers"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDTO"
        "401":
          description: "Unauthorized"
        "404":
          description: "User not found"
components:
  schemas:
    DemoDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
    UserDTO:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        userName:
          type: "string"
        email:
          type: "string"
          format: "email"
        password:
          type: "string"
          writeOnly: true
        firstName:
          type: "string"
        lastName:
          type: "string"
        birthDate:
          type: "string"
          format: "date"
        address:
          $ref: "#/components/schemas/AddressDTO"
    AddressDTO:
      properties:
        city:
          type: "string"
          description: "city of address"
    CreateUser:
      required:
      - "userName"
      - "email"
      - "password"
      type: "object"
      properties:
        userName:
          type: "string"
        email:
          type: "string"
          format: "email"
        password:
          type: "string"
          writeOnly: true
        firstName:
          type: "string"
        lastName:
          type: "string"
        birthDate:
          type: "string"
          format: "date"
        address:
          $ref: "#/components/schemas/AddressDTO"